requirements.txt
# Research Assistant Agent - Production Dependencies
# Optimized for Google Cloud Run deployment
# Last updated: November 2025

# ============================================================================
# Core AI/ML Framework
# ============================================================================
google-generativeai==0.3.2

# ============================================================================
# Environment & Configuration
# ============================================================================
python-dotenv==1.0.0

# ============================================================================
# Web Framework & API
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================================================
# HTTP Client & Networking
# ============================================================================
requests==2.31.0
httpx==0.25.2

# ============================================================================
# File Processing & Document Handling
# ============================================================================
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# ============================================================================
# Logging & Monitoring (Optional but Recommended)
# ============================================================================
colorlog==6.8.0

# ============================================================================
# Security & Performance (Production Enhancements)
# ============================================================================
# gunicorn==21.2.0  # Alternative to uvicorn for production
# python-json-logger==2.0.7  # Structured logging for Cloud Run

# ============================================================================
# Installation Instructions
# ============================================================================
# Local development:
#   pip install -r requirements.txt
#
# Cloud Run deployment:
#   Automatically installed during docker build
#
# Update packages:
#   pip install --upgrade -r requirements.txt
#
# Freeze current versions:
#   pip freeze > requirements-lock.txt